hostapd: fix upstream source URL
Paul Cercuei's mirror of hostapd for Realtek RTL8188 wifi chip is down.
I found a repo on github that seems to work (v2.0 release) at
https://github.com/jenssegers/RTL8188-hostapd

I am not sure of the long-term status or plans for that repo, so I have
forked it and used my github repo as the upstream URL for now.